execute on node, test passed

+++ lib/larrow/runner/manager.rb
@@ -9,26 +9,31 @@ module Larrow
       attr_accessor :vcs
       attr_accessor :app
       def initialize target_url
+        @state = :init
         self.vcs = Vcs.parse target_url
       end
 
-      def preload
-        self.vcs.load_manifest
-      end
-
       def go
-        self.app = Model::App.new vcs
+        preload
         allocate
         app.action
         release
       end
 
+      def preload
+        self.vcs.load_manifest
+        @state = :preload
+      end
+
       def allocate
-        app.assign node: Model::Node.new(*vm.create.first)
+        self.app = Model::App.new vcs
+        self.app.assign node: Model::Node.new(*vm.create.first)
+        @state = :allocate
       end
 
       def release
-        app.node.destroy
+        app.node.destroy if @state != :release
+        @state = :release
       end
     end
   end
